third_party/WebKit/LayoutTests/external/wpt/preload/onerror_event-expected.txt

CONSOLE WARNING: line 23: <link rel=preload> must have a valid `as` value
CONSOLE WARNING: line 24: <link rel=preload> must have a valid `as` value
CONSOLE WARNING: line 26: <link rel=preload> must have a valid `as` value
This is a testharness.js-based test.
FAIL Makes sure that preloaded resources trigger the onerror event assert_true: style triggered error event expected true got false
Harness: the test ran to completion.
